Implemented opcodes (e.g. 0xE0A1 - op0xF01E and op0xF055 - op0xF065).
In this commit, some of the opcodes required keyboard input so it was
time to map some of the GLFW inputs to suitable data structures to
be passed on a Chip8 instance. In order to do so, I had to introduce
a global variable in main.cpp because GLFW's input callback system
did not support passing in custom data to the callback in order to
allow setting the last key pressed on a specific Chip8 instance.
Hence, the solution was to keep track of the last key press for
each window, which is then passed to the Chip8 instance that was
associated with that window.

I changed some of the original data structures for holding key presses
to accomodate the state of the last key pressed. Although in the
future I think it would be a lot cleaner to introduce a struct
to separate the key states and the last key press as it would make
it more clearer.